Module trezor_protos::bitcoin

Modules

multisig_redeem_script_type

Nested message and enum types in MultisigRedeemScriptType.

tx_ack

Nested message and enum types in TxAck.

tx_ack_input

Nested message and enum types in TxAckInput.

tx_ack_output

Nested message and enum types in TxAckOutput.

tx_ack_payment_request

Nested message and enum types in TxAckPaymentRequest.

tx_ack_prev_extra_data

Nested message and enum types in TxAckPrevExtraData.

tx_ack_prev_input

Nested message and enum types in TxAckPrevInput.

tx_ack_prev_output

Nested message and enum types in TxAckPrevOutput.

tx_request

Nested message and enum types in TxRequest.

Structs

Address

AuthorizeCoinJoin

GetAddress

GetOwnershipId

GetOwnershipProof

GetPublicKey

MessageSignature

MultisigRedeemScriptType

OwnershipId

OwnershipProof

PrevInput

Data type for inputs of previous transactions.

PrevOutput

Data type for outputs of previous transactions. @embed

PrevTx

Data type for metadata about previous transaction which contains the UTXO being spent. @embed

PublicKey

SignMessage

SignTx

TxAck

TxAckInput

TxAckOutput

TxAckPaymentRequest

Data type of a payment request for a set of outputs. @next TxRequest

TxAckPrevExtraData

TxAckPrevInput

TxAckPrevMeta

TxAckPrevOutput

TxInput

Data type for transaction input to be signed.

TxOutput

Data type for transaction output to be signed. @embed

TxRequest

VerifyMessage

Enums

AmountUnit

DecredStakingSpendType

InputScriptType

OutputScriptType