Enum trezor_protos::MessageType

#[repr(i32)]
pub enum MessageType {

    Initialize,
    Ping,
    Success,
    Failure,
    ChangePin,
    WipeDevice,
    GetEntropy,
    Entropy,
    LoadDevice,
    ResetDevice,
    Features,
    PinMatrixRequest,
    PinMatrixAck,
    Cancel,
    LockDevice,
    ApplySettings,
    ButtonRequest,
    ButtonAck,
    ApplyFlags,
    GetNonce,
    Nonce,
    BackupDevice,
    EntropyRequest,
    EntropyAck,
    PassphraseRequest,
    PassphraseAck,
    RecoveryDevice,
    WordRequest,
    WordAck,
    GetFeatures,
    SdProtect,
    ChangeWipeCode,
    EndSession,
    DoPreauthorized,
    PreauthorizedRequest,
    CancelAuthorization,
    RebootToBootloader,
    GetFirmwareHash,
    FirmwareHash,
    GetFirmware,
    FirmwareChunk,
    FirmwareChunkAck,
    SetU2fCounter,
    GetNextU2fCounter,
    NextU2fCounter,
    DeprecatedPassphraseStateRequest,
    DeprecatedPassphraseStateAck,
    FirmwareErase,
    FirmwareUpload,
    FirmwareRequest,
    SelfTest,
    GetPublicKey,
    PublicKey,
    SignTx,
    TxRequest,
    TxAck,
    GetAddress,
    Address,
    TxAckPaymentRequest,
    SignMessage,
    VerifyMessage,
    MessageSignature,
    GetOwnershipId,
    OwnershipId,
    GetOwnershipProof,
    OwnershipProof,
    AuthorizeCoinJoin,
    CipherKeyValue,
    CipheredKeyValue,
    SignIdentity,
    SignedIdentity,
    GetEcdhSessionKey,
    EcdhSessionKey,
    CosiCommit,
    CosiCommitment,
    CosiSign,
    CosiSignature,
    DebugLinkDecision,
    DebugLinkGetState,
    DebugLinkState,
    DebugLinkStop,
    DebugLinkLog,
    DebugLinkMemoryRead,
    DebugLinkMemory,
    DebugLinkMemoryWrite,
    DebugLinkFlashErase,
    DebugLinkLayout,
    DebugLinkReseedRandom,
    DebugLinkRecordScreen,
    DebugLinkEraseSdCard,
    DebugLinkWatchLayout,
    EthereumGetPublicKey,
    EthereumPublicKey,
    EthereumGetAddress,
    EthereumAddress,
    EthereumSignTx,
    EthereumSignTxEip1559,
    EthereumTxRequest,
    EthereumTxAck,
    EthereumSignMessage,
    EthereumVerifyMessage,
    EthereumMessageSignature,
    EthereumSignTypedData,
    EthereumTypedDataStructRequest,
    EthereumTypedDataStructAck,
    EthereumTypedDataValueRequest,
    EthereumTypedDataValueAck,
    EthereumTypedDataSignature,
    EthereumSignTypedHash,
    NemGetAddress,
    NemAddress,
    NemSignTx,
    NemSignedTx,
    NemDecryptMessage,
    NemDecryptedMessage,
    TezosGetAddress,
    TezosAddress,
    TezosSignTx,
    TezosSignedTx,
    TezosGetPublicKey,
    TezosPublicKey,
    StellarSignTx,
    StellarTxOpRequest,
    StellarGetAddress,
    StellarAddress,
    StellarCreateAccountOp,
    StellarPaymentOp,
    StellarPathPaymentStrictReceiveOp,
    StellarManageSellOfferOp,
    StellarCreatePassiveSellOfferOp,
    StellarSetOptionsOp,
    StellarChangeTrustOp,
    StellarAllowTrustOp,
    StellarAccountMergeOp,
    StellarManageDataOp,
    StellarBumpSequenceOp,
    StellarManageBuyOfferOp,
    StellarPathPaymentStrictSendOp,
    StellarSignedTx,
    CardanoGetPublicKey,
    CardanoPublicKey,
    CardanoGetAddress,
    CardanoAddress,
    CardanoTxItemAck,
    CardanoTxAuxiliaryDataSupplement,
    CardanoTxWitnessRequest,
    CardanoTxWitnessResponse,
    CardanoTxHostAck,
    CardanoTxBodyHash,
    CardanoSignTxFinished,
    CardanoSignTxInit,
    CardanoTxInput,
    CardanoTxOutput,
    CardanoAssetGroup,
    CardanoToken,
    CardanoTxCertificate,
    CardanoTxWithdrawal,
    CardanoTxAuxiliaryData,
    CardanoPoolOwner,
    CardanoPoolRelayParameters,
    CardanoGetNativeScriptHash,
    CardanoNativeScriptHash,
    CardanoTxMint,
    CardanoTxCollateralInput,
    CardanoTxRequiredSigner,
    RippleGetAddress,
    RippleAddress,
    RippleSignTx,
    RippleSignedTx,
    MoneroTransactionInitRequest,
    MoneroTransactionInitAck,
    MoneroTransactionSetInputRequest,
    MoneroTransactionSetInputAck,
    MoneroTransactionInputViniRequest,
    MoneroTransactionInputViniAck,
    MoneroTransactionAllInputsSetRequest,
    MoneroTransactionAllInputsSetAck,
    MoneroTransactionSetOutputRequest,
    MoneroTransactionSetOutputAck,
    MoneroTransactionAllOutSetRequest,
    MoneroTransactionAllOutSetAck,
    MoneroTransactionSignInputRequest,
    MoneroTransactionSignInputAck,
    MoneroTransactionFinalRequest,
    MoneroTransactionFinalAck,
    MoneroKeyImageExportInitRequest,
    MoneroKeyImageExportInitAck,
    MoneroKeyImageSyncStepRequest,
    MoneroKeyImageSyncStepAck,
    MoneroKeyImageSyncFinalRequest,
    MoneroKeyImageSyncFinalAck,
    MoneroGetAddress,
    MoneroAddress,
    MoneroGetWatchKey,
    MoneroWatchKey,
    DebugMoneroDiagRequest,
    DebugMoneroDiagAck,
    MoneroGetTxKeyRequest,
    MoneroGetTxKeyAck,
    MoneroLiveRefreshStartRequest,
    MoneroLiveRefreshStartAck,
    MoneroLiveRefreshStepRequest,
    MoneroLiveRefreshStepAck,
    MoneroLiveRefreshFinalRequest,
    MoneroLiveRefreshFinalAck,
    EosGetPublicKey,
    EosPublicKey,
    EosSignTx,
    EosTxActionRequest,
    EosTxActionAck,
    EosSignedTx,
    BinanceGetAddress,
    BinanceAddress,
    BinanceGetPublicKey,
    BinancePublicKey,
    BinanceSignTx,
    BinanceTxRequest,
    BinanceTransferMsg,
    BinanceOrderMsg,
    BinanceCancelMsg,
    BinanceSignedTx,
    WebAuthnListResidentCredentials,
    WebAuthnCredentials,
    WebAuthnAddResidentCredential,
    WebAuthnRemoveResidentCredential,

}

Mapping between Trezor wire identifier (uint) and a protobuf message

Variants

Initialize

Management

Ping

Success

Failure

ChangePin

WipeDevice

GetEntropy

Entropy

LoadDevice

ResetDevice

Features

PinMatrixRequest

PinMatrixAck

Cancel

LockDevice

ApplySettings

ButtonRequest

ButtonAck

ApplyFlags

GetNonce

Nonce

BackupDevice

EntropyRequest

EntropyAck

PassphraseRequest

PassphraseAck

RecoveryDevice

WordRequest

WordAck

GetFeatures

SdProtect

ChangeWipeCode

EndSession

DoPreauthorized

PreauthorizedRequest

CancelAuthorization

RebootToBootloader

GetFirmwareHash

FirmwareHash

GetFirmware

FirmwareChunk

FirmwareChunkAck

SetU2fCounter

GetNextU2fCounter

NextU2fCounter

DeprecatedPassphraseStateRequest

Deprecated messages, kept for protobuf compatibility. Both are marked wire_out so that we don’t need to implement incoming handler for legacy

DeprecatedPassphraseStateAck

FirmwareErase

Bootloader

FirmwareUpload

FirmwareRequest

SelfTest

GetPublicKey

Bitcoin

PublicKey

SignTx

TxRequest

TxAck

GetAddress

Address

TxAckPaymentRequest

SignMessage

VerifyMessage

MessageSignature

GetOwnershipId

OwnershipId

GetOwnershipProof

OwnershipProof

AuthorizeCoinJoin

CipherKeyValue

Crypto

CipheredKeyValue

SignIdentity

SignedIdentity

GetEcdhSessionKey

EcdhSessionKey

CosiCommit

CosiCommitment

CosiSign

CosiSignature

DebugLinkDecision

Debug

DebugLinkGetState

DebugLinkState

DebugLinkStop

DebugLinkLog

DebugLinkMemoryRead

DebugLinkMemory

DebugLinkMemoryWrite

DebugLinkFlashErase

DebugLinkLayout

DebugLinkReseedRandom

DebugLinkRecordScreen

DebugLinkEraseSdCard

DebugLinkWatchLayout

EthereumGetPublicKey

Ethereum

EthereumPublicKey

EthereumGetAddress

EthereumAddress

EthereumSignTx

EthereumSignTxEip1559

EthereumTxRequest

EthereumTxAck

EthereumSignMessage

EthereumVerifyMessage

EthereumMessageSignature

EthereumSignTypedData

EthereumTypedDataStructRequest

EthereumTypedDataStructAck

EthereumTypedDataValueRequest

EthereumTypedDataValueAck

EthereumTypedDataSignature

EthereumSignTypedHash

NemGetAddress

NEM

NemAddress

NemSignTx

NemSignedTx

NemDecryptMessage

NemDecryptedMessage

TezosGetAddress

Tezos

TezosAddress

TezosSignTx

TezosSignedTx

TezosGetPublicKey

TezosPublicKey

StellarSignTx

Stellar

StellarTxOpRequest

StellarGetAddress

StellarAddress

StellarCreateAccountOp

StellarPaymentOp

StellarPathPaymentStrictReceiveOp

StellarManageSellOfferOp

StellarCreatePassiveSellOfferOp

StellarSetOptionsOp

StellarChangeTrustOp

StellarAllowTrustOp

StellarAccountMergeOp

StellarManageDataOp

omitted: StellarInflationOp is not a supported operation, would be 219

StellarBumpSequenceOp

StellarManageBuyOfferOp

StellarPathPaymentStrictSendOp

StellarSignedTx

CardanoGetPublicKey

CardanoPublicKey

CardanoGetAddress

CardanoAddress

CardanoTxItemAck

CardanoTxAuxiliaryDataSupplement

CardanoTxWitnessRequest

CardanoTxWitnessResponse

CardanoTxHostAck

CardanoTxBodyHash

CardanoSignTxFinished

CardanoSignTxInit

CardanoTxInput

CardanoTxOutput

CardanoAssetGroup

CardanoToken

CardanoTxCertificate

CardanoTxWithdrawal

CardanoTxAuxiliaryData

CardanoPoolOwner

CardanoPoolRelayParameters

CardanoGetNativeScriptHash

CardanoNativeScriptHash

CardanoTxMint

CardanoTxCollateralInput

CardanoTxRequiredSigner

RippleGetAddress

Ripple

RippleAddress

RippleSignTx

RippleSignedTx

MoneroTransactionInitRequest

Monero

MoneroTransactionInitAck

MoneroTransactionSetInputRequest

MoneroTransactionSetInputAck

MoneroTransactionInputViniRequest

MoneroTransactionInputViniAck

MoneroTransactionAllInputsSetRequest

MoneroTransactionAllInputsSetAck

MoneroTransactionSetOutputRequest

MoneroTransactionSetOutputAck

MoneroTransactionAllOutSetRequest

MoneroTransactionAllOutSetAck

MoneroTransactionSignInputRequest

MoneroTransactionSignInputAck

MoneroTransactionFinalRequest

MoneroTransactionFinalAck

MoneroKeyImageExportInitRequest

MoneroKeyImageExportInitAck

MoneroKeyImageSyncStepRequest

MoneroKeyImageSyncStepAck

MoneroKeyImageSyncFinalRequest

MoneroKeyImageSyncFinalAck

MoneroGetAddress

MoneroAddress

MoneroGetWatchKey

MoneroWatchKey

DebugMoneroDiagRequest

DebugMoneroDiagAck

MoneroGetTxKeyRequest

MoneroGetTxKeyAck

MoneroLiveRefreshStartRequest

MoneroLiveRefreshStartAck

MoneroLiveRefreshStepRequest

MoneroLiveRefreshStepAck

MoneroLiveRefreshFinalRequest

MoneroLiveRefreshFinalAck

EosGetPublicKey

EOS

EosPublicKey

EosSignTx

EosTxActionRequest

EosTxActionAck

EosSignedTx

BinanceGetAddress

Binance

BinanceAddress

BinanceGetPublicKey

BinancePublicKey

BinanceSignTx

BinanceTxRequest

BinanceTransferMsg

BinanceOrderMsg

BinanceCancelMsg

BinanceSignedTx

WebAuthnListResidentCredentials

WebAuthn

WebAuthnCredentials

WebAuthnAddResidentCredential

WebAuthnRemoveResidentCredential

Implementations

impl MessageType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of MessageType.

pub fn from_i32(value: i32) -> Option<MessageType>

Converts an i32 to a MessageType, or None if value is not a valid variant.