Struct trezor_protos::management::Features

pub struct Features {

    pub vendor: Option<String>,
    pub major_version: u32,
    pub minor_version: u32,
    pub patch_version: u32,
    pub bootloader_mode: Option<bool>,
    pub device_id: Option<String>,
    pub pin_protection: Option<bool>,
    pub passphrase_protection: Option<bool>,
    pub language: Option<String>,
    pub label: Option<String>,
    pub initialized: Option<bool>,
    pub revision: Option<Vec<u8>>,
    pub bootloader_hash: Option<Vec<u8>>,
    pub imported: Option<bool>,
    pub unlocked: Option<bool>,
    pub passphrase_cached: Option<bool>,
    pub firmware_present: Option<bool>,
    pub needs_backup: Option<bool>,
    pub flags: Option<u32>,
    pub model: Option<String>,
    pub fw_major: Option<u32>,
    pub fw_minor: Option<u32>,
    pub fw_patch: Option<u32>,
    pub fw_vendor: Option<String>,
    pub unfinished_backup: Option<bool>,
    pub no_backup: Option<bool>,
    pub recovery_mode: Option<bool>,
    pub capabilities: Vec<i32>,
    pub backup_type: Option<i32>,
    pub sd_card_present: Option<bool>,
    pub sd_protection: Option<bool>,
    pub wipe_code_protection: Option<bool>,
    pub session_id: Option<Vec<u8>>,
    pub passphrase_always_on_device: Option<bool>,
    pub safety_checks: Option<i32>,
    pub auto_lock_delay_ms: Option<u32>,
    pub display_rotation: Option<u32>,
    pub experimental_features: Option<bool>,

}

Response: Reports various information about the device @end

Fields
vendor: Option<String>

name of the manufacturer, e.g. “trezor.io”

major_version: u32

major version of the firmware/bootloader, e.g. 1

minor_version: u32

minor version of the firmware/bootloader, e.g. 0

patch_version: u32

patch version of the firmware/bootloader, e.g. 0

bootloader_mode: Option<bool>

is device in bootloader mode?

device_id: Option<String>

device’s unique identifier

pin_protection: Option<bool>

is device protected by PIN?

passphrase_protection: Option<bool>

is node/mnemonic encrypted using passphrase?

language: Option<String>

device language

label: Option<String>

device description label

initialized: Option<bool>

does device contain seed?

revision: Option<Vec<u8>>

SCM revision of firmware

bootloader_hash: Option<Vec<u8>>

hash of the bootloader

imported: Option<bool>

was storage imported from an external source?

unlocked: Option<bool>

is the device unlocked? called “pin_cached” previously

passphrase_cached: Option<bool>

👎 Deprecated

is passphrase already cached in session?

firmware_present: Option<bool>

is valid firmware loaded?

needs_backup: Option<bool>

does storage need backup? (equals to Storage.needs_backup)

flags: Option<u32>

device flags (equals to Storage.flags)

model: Option<String>

device hardware model

fw_major: Option<u32>

reported firmware version if in bootloader mode

fw_minor: Option<u32>

reported firmware version if in bootloader mode

fw_patch: Option<u32>

reported firmware version if in bootloader mode

fw_vendor: Option<String>

reported firmware vendor if in bootloader mode

unfinished_backup: Option<bool>

optional bytes fw_vendor_keys = 26; // obsoleted, use fw_vendor

report unfinished backup (equals to Storage.unfinished_backup)

no_backup: Option<bool>

report no backup (equals to Storage.no_backup)

recovery_mode: Option<bool>

is recovery mode in progress

capabilities: Vec<i32>

list of supported capabilities

backup_type: Option<i32>

type of device backup (BIP-39 / SLIP-39 basic / SLIP-39 advanced)

sd_card_present: Option<bool>

is SD card present

sd_protection: Option<bool>

is SD Protect enabled

wipe_code_protection: Option<bool>

is wipe code protection enabled

session_id: Option<Vec<u8>>passphrase_always_on_device: Option<bool>

device enforces passphrase entry on Trezor

safety_checks: Option<i32>

safety check level, set to Prompt to limit path namespace enforcement

auto_lock_delay_ms: Option<u32>

number of milliseconds after which the device locks itself

display_rotation: Option<u32>

in degrees from North

experimental_features: Option<bool>

are experimental message types enabled?

Implementations

impl Features

pub fn vendor(&self) -> &str

Returns the value of vendor, or the default value if vendor is unset.

pub fn bootloader_mode(&self) -> bool

Returns the value of bootloader_mode, or the default value if bootloader_mode is unset.

pub fn device_id(&self) -> &str

Returns the value of device_id, or the default value if device_id is unset.

pub fn pin_protection(&self) -> bool

Returns the value of pin_protection, or the default value if pin_protection is unset.

pub fn passphrase_protection(&self) -> bool

Returns the value of passphrase_protection, or the default value if passphrase_protection is unset.

pub fn language(&self) -> &str

Returns the value of language, or the default value if language is unset.

pub fn label(&self) -> &str

Returns the value of label, or the default value if label is unset.

pub fn initialized(&self) -> bool

Returns the value of initialized, or the default value if initialized is unset.

pub fn revision(&self) -> &[u8]

Returns the value of revision, or the default value if revision is unset.

pub fn bootloader_hash(&self) -> &[u8]

Returns the value of bootloader_hash, or the default value if bootloader_hash is unset.

pub fn imported(&self) -> bool

Returns the value of imported, or the default value if imported is unset.

pub fn unlocked(&self) -> bool

Returns the value of unlocked, or the default value if unlocked is unset.

pub fn passphrase_cached(&self) -> bool

Returns the value of passphrase_cached, or the default value if passphrase_cached is unset.

pub fn firmware_present(&self) -> bool

Returns the value of firmware_present, or the default value if firmware_present is unset.

pub fn needs_backup(&self) -> bool

Returns the value of needs_backup, or the default value if needs_backup is unset.

pub fn flags(&self) -> u32

Returns the value of flags, or the default value if flags is unset.

pub fn model(&self) -> &str

Returns the value of model, or the default value if model is unset.

pub fn fw_major(&self) -> u32

Returns the value of fw_major, or the default value if fw_major is unset.

pub fn fw_minor(&self) -> u32

Returns the value of fw_minor, or the default value if fw_minor is unset.

pub fn fw_patch(&self) -> u32

Returns the value of fw_patch, or the default value if fw_patch is unset.

pub fn fw_vendor(&self) -> &str

Returns the value of fw_vendor, or the default value if fw_vendor is unset.

pub fn unfinished_backup(&self) -> bool

Returns the value of unfinished_backup, or the default value if unfinished_backup is unset.

pub fn no_backup(&self) -> bool

Returns the value of no_backup, or the default value if no_backup is unset.

pub fn recovery_mode(&self) -> bool

Returns the value of recovery_mode, or the default value if recovery_mode is unset.

pub fn capabilities(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<Capability>>

Returns an iterator which yields the valid enum values contained in capabilities.

pub fn push_capabilities(&mut self, value: Capability)

Appends the provided enum value to capabilities.

pub fn backup_type(&self) -> BackupType

Returns the enum value of backup_type, or the default if the field is unset or set to an invalid enum value.

pub fn set_backup_type(&mut self, value: BackupType)

Sets backup_type to the provided enum value.

pub fn sd_card_present(&self) -> bool

Returns the value of sd_card_present, or the default value if sd_card_present is unset.

pub fn sd_protection(&self) -> bool

Returns the value of sd_protection, or the default value if sd_protection is unset.

pub fn wipe_code_protection(&self) -> bool

Returns the value of wipe_code_protection, or the default value if wipe_code_protection is unset.

pub fn session_id(&self) -> &[u8]

Returns the value of session_id, or the default value if session_id is unset.

pub fn passphrase_always_on_device(&self) -> bool

Returns the value of passphrase_always_on_device, or the default value if passphrase_always_on_device is unset.

pub fn safety_checks(&self) -> SafetyCheckLevel

Returns the enum value of safety_checks, or the default if the field is unset or set to an invalid enum value.

pub fn set_safety_checks(&mut self, value: SafetyCheckLevel)

Sets safety_checks to the provided enum value.

pub fn auto_lock_delay_ms(&self) -> u32

Returns the value of auto_lock_delay_ms, or the default value if auto_lock_delay_ms is unset.

pub fn display_rotation(&self) -> u32

Returns the value of display_rotation, or the default value if display_rotation is unset.

pub fn experimental_features(&self) -> bool

Returns the value of experimental_features, or the default value if experimental_features is unset.