Struct trezor_protos::nem::nem_sign_tx::nem_mosaic_creation::NemMosaicDefinition

pub struct NemMosaicDefinition {

    pub name: Option<String>,
    pub ticker: Option<String>,
    pub namespace: String,
    pub mosaic: String,
    pub divisibility: Option<u32>,
    pub levy: Option<i32>,
    pub fee: Option<u64>,
    pub levy_address: Option<String>,
    pub levy_namespace: Option<String>,
    pub levy_mosaic: Option<String>,
    pub supply: Option<u64>,
    pub mutable_supply: Option<bool>,
    pub transferable: Option<bool>,
    pub description: String,
    pub networks: Vec<u32>,

}

Structure representing a mosaic definition

Fields
name: Option<String>

User-friendly name of the mosaic (for whitelisted mosaics)

ticker: Option<String>

Ticker of the mosaic (for whitelisted mosaics)

namespace: String

Fully qualified name of the namespace

mosaic: String

Name of the mosaic definition

divisibility: Option<u32>

Number of decimal places that a mosaic can be divided into

levy: Option<i32>

Levy type

fee: Option<u64>

Levy fee (interpretation depends on levy type)

levy_address: Option<String>

Levy address

levy_namespace: Option<String>

Fully qualified name of the namespace of the levy mosaic

levy_mosaic: Option<String>

Name of the levy mosaic

supply: Option<u64>

Initial supply to create, always given in entire units

mutable_supply: Option<bool>

Mutable supply

transferable: Option<bool>

Mosaic allows transfers among accounts other than the creator

description: String

Mosaic description

networks: Vec<u32>

Networks that the mosaic is valid on (for whitelisted mosaics)

Implementations

impl NemMosaicDefinition

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn ticker(&self) -> &str

Returns the value of ticker, or the default value if ticker is unset.

pub fn divisibility(&self) -> u32

Returns the value of divisibility, or the default value if divisibility is unset.

pub fn levy(&self) -> NemMosaicLevy

Returns the enum value of levy, or the default if the field is unset or set to an invalid enum value.

pub fn set_levy(&mut self, value: NemMosaicLevy)

Sets levy to the provided enum value.

pub fn fee(&self) -> u64

Returns the value of fee, or the default value if fee is unset.

pub fn levy_address(&self) -> &str

Returns the value of levy_address, or the default value if levy_address is unset.

pub fn levy_namespace(&self) -> &str

Returns the value of levy_namespace, or the default value if levy_namespace is unset.

pub fn levy_mosaic(&self) -> &str

Returns the value of levy_mosaic, or the default value if levy_mosaic is unset.

pub fn supply(&self) -> u64

Returns the value of supply, or the default value if supply is unset.

pub fn mutable_supply(&self) -> bool

Returns the value of mutable_supply, or the default value if mutable_supply is unset.

pub fn transferable(&self) -> bool

Returns the value of transferable, or the default value if transferable is unset.